Make churn job into a view.
